+++ src/main/kotlin/com/rankweis/uppercut/karate/run/KarateRunConfiguration.java
@@ -20,7 +20,6 @@ import com.intellij.execution.target.TargetEnvironmentConfiguration;
 import com.intellij.execution.testframework.sm.SMTestRunnerConnectionUtil;
 import com.intellij.execution.testframework.sm.runner.ui.SMTRunnerConsoleView;
 import com.intellij.execution.ui.ConsoleView;
-import com.intellij.ide.util.PropertiesComponent;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.application.ModalityState;
 import com.intellij.openapi.application.ReadAction;
@@ -28,6 +27,7 @@ import com.intellij.openapi.options.SettingsEditor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.PathUtil;
+import com.rankweis.uppercut.settings.KarateSettingsState;
 import com.rankweis.uppercut.testrunner.KarateTestRunner;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -65,7 +65,7 @@ public class KarateRunConfiguration extends ApplicationConfiguration implements 
   @Getter @Setter private String tag;
   @Getter @Setter private String path;
   @Getter @Setter private PreferredTest preferredTest = PreferredTest.WHOLE_FILE;
-  @Getter @Setter private String parallelism = "5";
+  @Getter @Setter private String parallelism;
   @Getter @Setter private boolean allInFolderAreFeature = false;
   private String environment;
 
@@ -209,7 +209,8 @@ public class KarateRunConfiguration extends ApplicationConfiguration implements 
     element.setAttribute("tag", Optional.ofNullable(tag).orElse(""));
     element.setAttribute("path", Optional.ofNullable(path).orElse(""));
     element.setAttribute("preferredTest", preferredTest.name);
-    element.setAttribute("parallelism", Optional.ofNullable(parallelism).orElse("1"));
+    element.setAttribute("parallelism", Optional.ofNullable(parallelism).orElse(
+      Optional.ofNullable(KarateSettingsState.getInstance().getDefaultParallelism()).map(String::valueOf).orElse("1")));
     element.setAttribute("allInFolderAreFeature", String.valueOf(allInFolderAreFeature));
     element.setAttribute("relPath", Optional.ofNullable(relPath).orElse(""));
     super.writeExternal(element);
@@ -239,7 +240,12 @@ public class KarateRunConfiguration extends ApplicationConfiguration implements 
 
   public String getEnv() {
     return StringUtil.isEmpty(environment) ?
-      PropertiesComponent.getInstance().getValue("uppercut.settings.defaultEnvironment") : environment;
+       String.valueOf(KarateSettingsState.getInstance().getDefaultEnvironment()) : environment;
+  }
+
+  public String getParallelism() {
+    return StringUtil.isEmpty(parallelism) ?
+       String.valueOf(KarateSettingsState.getInstance().getDefaultParallelism()) : parallelism;
   }
 
   public void setEnv(String environment) {